Serialization - [Fix] Ensure out-of-order propertytable array elements are read
PropertyTables serialize arrays into a number of separate elements with names such as children::0, children::1, children::2 etc.

Previously, the property table would fail to find all the elements unless they were located in the properties file in the correct order.

This change ensures the entire properties list is scanned for each element, so out-of-order elements are not missed.